% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_function.R
\name{get_ahrefs_data}
\alias{get_ahrefs_data}
\title{Get Data From the AHREF's}
\usage{
get_ahrefs_data(target, token, from, mode, select = NULL, where = NULL,
  having = NULL, order_by = NULL, offset = 0, limit = 1000)
}
\arguments{
\item{target}{Aim of a request: a domain, a directory or a URL}

\item{token}{Authentication token}

\item{from}{Table to select data from}

\item{mode}{Mode of operation: exact, domain, subdomains or prefix}

\item{select}{A character vector of columns to select in 'from' table.}

\item{where}{"Where" condition to satisfy (see documentation for how to build where conditions)}

\item{having}{"Having" condition to satisfy (see documentation for how to build a having conditions)}

\item{order_by}{List of columns to sort on (see documentation for how to build order_by conitions)}

\item{offset}{Position from which to start returning data}

\item{limit}{Number of results to return}
}
\value{
An object of class 'ahrefs_api'
}
\description{
Get data from the AHREF's API.
For more information see the online documentation: https://ahrefs.com/api/documentation.
}
\examples{
\dontrun{
ahrefs_key <- '123456789'
get_ahrefs_data('mhairihmcneill.com', ahrefs_key, from = 'ahrefs_rank', mode = 'domain')
get_ahrefs_data('mhairihmcneill.com/blog', ahrefs_key, from = 'metrics_extended', mode = 'exact')
get_ahrefs_data('mhairihmcneill.com/blog', ahrefs_key, from = 'refdomains', mode = 'prefix', limit = 10)
}
}

