% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_functions.R
\name{get_domain_rating}
\alias{get_domain_rating}
\title{Get Domain Rating}
\usage{
get_domain_rating(url, api_key)
}
\arguments{
\item{url}{URL we want the rank for}

\item{api_key}{API key (also called token)}
}
\value{
A single number
}
\description{
Depreciated: will be removed soon
}
\examples{
\dontrun{
get_domain_rating('google.com', ahrefs_key)
}
}

