% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_metrics}
\alias{get_metrics}
\title{Get URL metrics}
\usage{
get_metrics(url, api_key = "1f5295f58ebd7f5622c803adf3bf40ce16139a6f")
}
\arguments{
\item{url}{URL we metrics for}

\item{api_key}{API key (also called token)}
}
\value{
A list
}
\description{
Get URL metrics
}
\examples{
\dontrun{
get_metrics('google.com', '1f5295f58ebd7f5622c803adf3bf40ce16139a6f')
}
}

