% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_functions.R
\name{get_rank}
\alias{get_rank}
\title{Get AHREF rank}
\usage{
get_rank(url, api_key)
}
\arguments{
\item{url}{URL we want the rank for}

\item{api_key}{API key (also called token)}
}
\value{
A single number
}
\description{
Depreciated: will be removed soon
}
\examples{
\dontrun{
get_rank('google.com', ahrefs_key)
}
}

