% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_functions.R
\name{get_refering_domains}
\alias{get_refering_domains}
\title{Get Refering Domains}
\usage{
get_refering_domains(url, api_key, limit = 10)
}
\arguments{
\item{url}{URL we want the domains for}

\item{api_key}{API key (also called token)}
}
\value{
A list: first element has info on the refering domains, second on overall statistics
}
\description{
Depreciated: will be removed soon
}
\examples{
\dontrun{
get_refering_domains('google.com', ahrefs_key)
}
}

