% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make_condition}
\alias{make_condition}
\title{Make a condition.}
\usage{
make_condition(col, operator, value)
}
\arguments{
\item{col}{List of columns to operate on}

\item{operator}{List of operators to use. Possible operators are: '=', '<>', '<', '<=', '>', '>=', 'subdomain', 'word', 'substring'.}

\item{value}{List of values for operator.}
}
\value{
A single element URL encoded character vector.
}
\description{
Make a contition to use as 'having' or 'where' paramater.
}
\examples{
make_condition('title', 'word', 'the')
make_condition(c('date', 'date'), c('=', '='), c('2016-10-01', '2016-10-02'))

}

