% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make_order}
\alias{make_order}
\title{Make an order.}
\usage{
make_order(col, direction = "asc")
}
\arguments{
\item{col}{Columns for ordering}

\item{direction}{Character vector of either 'asc' for accending, or 'desc' for decending.}
}
\value{
A single element URL encoded character vector.
}
\description{
Make an order, to use in an order parameter.
}
\examples{

make_order('date')
make_order(c('year', 'month'), c('desc', 'desc'))
}

