% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_domain_rating}
\alias{get_domain_rating}
\title{Get Domain Rating}
\usage{
get_domain_rating(url, api_key)
}
\arguments{
\item{url}{URL we want the rank for}

\item{api_key}{API key (also called token)}
}
\value{
A single number
}
\description{
Get Domain Rating
}
\examples{
\dontrun{
get_domain_rating('google.com', '1f5295f58ebd7f5622c803adf3bf40ce16139a6f')
}
}

