% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_functions.R
\name{get_metrics}
\alias{get_metrics}
\title{Get URL metrics}
\usage{
get_metrics(url, api_key)
}
\arguments{
\item{url}{URL we metrics for}

\item{api_key}{API key (also called token)}
}
\value{
A list
}
\description{
Depreciated: will be removed soon
}
\examples{
\dontrun{
get_metrics('google.com', ahrefs_key)
}
}

