% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_rank}
\alias{get_rank}
\title{Get AHREF rank}
\usage{
get_rank(url, api_key)
}
\arguments{
\item{url}{URL we want the rank for}

\item{api_key}{API key (also called token)}
}
\value{
A single number
}
\description{
Get AHREF rank
}
\examples{
\dontrun{
get_rank('google.com', '1f5295f58ebd7f5622c803adf3bf40ce16139a6f')
}
}

