% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_refering_domains}
\alias{get_refering_domains}
\title{Get Refering Domains}
\usage{
get_refering_domains(url, api_key, limit = 10)
}
\arguments{
\item{url}{URL we want the domains for}

\item{api_key}{API key (also called token)}
}
\value{
A list: first element has info on the refering domains, second on overall statistics
}
\description{
Get Refering Domains
}
\examples{
\dontrun{
get_refering_domains('google.com', '1f5295f58ebd7f5622c803adf3bf40ce16139a6f')
}
}

